+++ src/tag/BlockTag.php
@@ -217,14 +217,6 @@ enum BlockTag: string
             throw new InvalidArgumentException(Message::EMPTY_TAG_NAME->getMessage());
         }
 
-        $tag = strtolower($tag);
-
-        foreach (self::cases() as $case) {
-            if ($case->value === $tag) {
-                return true;
-            }
-        }
-
-        return false;
+        return self::tryFrom(strtolower($tag)) !== null;
     }
 }